Implementado o sistema de saida automatica

+++ tserver.c
@@ -58,7 +58,7 @@
      	//se sockfd < 0 (ie -1) quer dizer que o socket nao foi criado 
      	if(sockfd < 0)
      	{
-     		printf("Erro ao abrir socket");
+     		printf("Erro ao abrir socket\n");
      		exit(1);
      	}
 
@@ -84,7 +84,7 @@
      	//leva o discritor do socket, o endereco para unir e o tamanho do mesmo; 
      	if(bind(sockfd,(struct sockaddr*)&serv_addr, sizeof(serv_addr)) < 0)
      	{
-     		printf("Erro no binding");
+     		printf("Erro no binding\n");
      		exit(1);
      	}
 
@@ -103,18 +103,18 @@
      	newsockfd = accept(sockfd, (struct sockaddr*)&cli_addr, &clilen);
      	if(newsockfd < 0)
      	{
-     		printf("Erro na conexao");
+     		printf("Erro na conexao\n");
      		exit(1);
      	}
-		printf("conexao estabelecida\n");
+		printf(">Conexao Estabelecida\n");
      	//para mandar coisas para o cliente, usamos write
      	//colocamos o descritor de arquivo do socket, o buffer (nesse caso string) de conteudo e o tamanho.
-     	n = write(newsockfd,"conexao estabelecida\n",21);
+     	n = write(newsockfd,"Conexao Estabelecida\n",21);
      	
      	//n guarda quantos caracteres foram mandados, se nao mandou nada = erro
      	if(n < 0)
      	{
-     		printf("Erro ao mandar texto");
+     		printf("Erro ao mandar texto\n");
      		exit(1);
      	}
 
@@ -123,7 +123,7 @@
 		pid_t pid = fork();
 
 		if(pid < 0 ){
-			perror("Fork Falhou");
+			perror("Fork Falhou\n");
 		}
 
 		//nesse caso, vamos colocar o escrever mensagens no child, nao acho que muda muito	
@@ -140,21 +140,20 @@
                 if(strcmp(buffer,empty) == 0)
                 {
 
-                }
-            
+                }           
 				
 
 				if(strcmp(sair,buffer) == 0) //se acabou a conversa
 				{
 					n = write(newsockfd,"Conversa Terminada X.X\n",23);
-					exit(0);
+					exit(2); //envia para o processo pai que terminou
 				}
 
 
 				n = write(newsockfd,buffer,sizeof(buffer));
 				if(n < 0)
 				{
-					printf("Erro ao escrever mensagem, terminando chatroom");
+					printf("Erro ao escrever mensagem, terminando chatroom\n");
 					exit(1);
 				}
 			}
@@ -170,12 +169,23 @@
 			{
 				bzero(buffer2,sizeof(buffer));
 				n = read(newsockfd,buffer2,sizeof(buffer2)); //lemos do buffer
-				if(n < 0)
-				{
-					printf("Erro em receber mensagem, ou acabou a transmissao. Saindo...\n");
+
+                //Caso o servidor queira sair da conversa
+                int pidfilho, status;
+                pidfilho = waitpid(pid, &status, WNOHANG); //verifica se o processo filho mandou algo
+                if(pidfilho < 0){
+                    perror("waitpid\n");
+                }
+                if(pidfilho > 0){
+                    exit(0);
+                }
+
+				if(n < 0){
+				    printf("Erro em receber mensagem, ou acabou a transmissao.\nSaindo...\n");
                     kill(pid, SIGTERM); //mata processo filho
 					exit(1);
 				}
+
 				char* empty = "";
                 if(strcmp(buffer2,empty) == 0)
                 {
@@ -183,7 +193,7 @@
                 }
 				else if(strcmp(buffer2,"Conversa Terminada X.X\n") == 0)
                 {
-                    printf("O outro usuario terminou a conversa\n Saindo...\n");
+                    printf(" O outro usuario terminou a conversa\n Saindo...\n");
                     kill(pid, SIGTERM);//mata processo filho
                     exit(0);
                 }
